requirements.txt
fastapi
uvicorn
langgraph
langchain-core
langchain_openai
langchain_experimental
langchain_text_splitters
langchain_huggingface
langchain_chroma
langchain_community
chromadb
unstructured
python-dotenv
tavily-python
ollama
pdfplumber
line-bot-sdk
requests